requirements.txt
pyyaml
argparse
cerberus
setuptools
